build: doing some general code cleanup
