+++ src/features/cv/components/cv-content-panel/cv-content-panel.tsx
