+++ src/features/cv/components/cv-education-item/cv-education-item.tsx
