+++ src/features/cv/components/cv-education-panel/cv-education-panel.tsx
