+++ src/features/cv/components/cv-skill-category/cv-skill-category.spec.tsx
