+++ src/features/cv/components/cv-skills-panel/cv-skills-panel.tsx
