+++ src/features/github/components/github-repositories/github-repositories.tsx
@@ -1,7 +1,7 @@
 import Container from '@/components/common/layout/container/container';
 import FlexContainer from '@/components/common/layout/flex-container/flex-container';
 import FlexItem from '@/components/common/layout/flex-item/flex-item';
-import { GetUserRepositoriesResponse } from '@/services/github/responses/GetUserRepositoriesResponse';
+import { GetUserRepositoriesResponse } from '@/features/github/services/responses/GetUserRepositoriesResponse';
 import { motion } from 'framer-motion';
 import React from 'react';
 import GithubRepositoryItem from '../github-repository-item/github-repository-item';
