--- src/styles/partials/_colours.scss
@@ -1,7 +1,0 @@
-$background: #121215;
-$accentBlue: #90fcf9;
-$darkerBlue: #4d7c8a;
-$lightGrey: #dbd3d8;
-$copper: #c57b57;
-$fontMain: #fafafa;
-$fontDark: #121215;
