+++ src/styles/partials/_typography.scss
@@ -3,7 +3,7 @@
 html,
 body {
   font-size: 18px;
-  font-family: 'Poppins', Arial, Helvetica, sans-serif;
+  @apply font-sans;
 }
 
 h1,
@@ -12,8 +12,7 @@ h3,
 h4,
 h5,
 h6 {
-  font-family: 'Source Code Pro', monospace;
-  font-weight: 700;
+  @apply font-mono font-bold;
 }
 
 h1 {
